Migrated org class code into org.py
